added the test cases in test_query.py
