+++ sympy/core/tests/test_assumptions.py
@@ -1,4 +1,3 @@
-from sympy.assumptions import ask, Q
 from sympy.core.mod import Mod
 from sympy.core.numbers import (I, oo, pi)
 from sympy.functions.combinatorial.factorials import factorial
@@ -1226,20 +1225,6 @@ def test_issue_21651():
     k = Symbol('k', positive=True, integer=True)
     exp = 2*2**(-k)
     assert exp.is_integer is None
-
-
-def test_issue_27449():
-    assert ask(Q.imaginary(I * oo)) is False
-    assert ask(Q.imaginary(oo)) is False
-    assert ask(Q.imaginary(2*I + oo)) is False
-    assert ask(Q.imaginary(3*I * oo)) is False
-    assert ask(Q.imaginary((I*oo)**2)) is False
-
-    i = Symbol('i', imaginary=True)
-    r = Symbol('r', real=True)
-    assert ask(Q.imaginary(r * i * oo)) is False
-    assert ask(Q.imaginary(i + r + oo)) is False
-    assert ask(Q.imaginary((i*oo)**r)) is False
 
 
 def test_assumptions_copy():
